+++ src-opam/distro.mli
@@ -114,7 +114,8 @@ type distro =
     | `V33
     | `V34
     | `V35
-    | `V36 ]
+    | `V36
+    | `V37 ]
   | `OracleLinux of [ `V7 | `V8 ]
   | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `V15_2 | `V15_3 ]
   | `Ubuntu of
@@ -178,6 +179,7 @@ type t =
     | `V34
     | `V35
     | `V36
+    | `V37
     | `Latest ]
   | `OracleLinux of [ `V7 | `V8 | `Latest ]
   | `OpenSUSE of